chore: reorder files and functions
