+++ Makefile
@@ -6,7 +6,7 @@
 #    By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/08/01 15:11:43 by pmarkaid          #+#    #+#              #
-#    Updated: 2024/08/18 14:23:37 by pmarkaid         ###   ########.fr        #
+#    Updated: 2024/08/21 19:00:37 by pmarkaid         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -33,8 +33,9 @@ SRCS_FILES = \
 	execution_utils.c \
 	error.c \
 	dup.c \
-	clean.c \
-	clean_utils.c
+	clean_utils.c \
+	expand.c \
+	split_args.c
 
 SRC_DIR = src/
 SRCS = $(addprefix $(SRC_DIR), $(SRCS_FILES))
